refactor(user): 유스케이스와 도메인 서비스 분리
